+++ test/support/kinetic_ecto_test_migration.ex
@@ -1,0 +1,14 @@
+defmodule KineticEcto.TestMigration do
+  @moduledoc false
+
+  use Ecto.Migration
+
+  def change do
+    create table(:images) do
+      add :url, :string
+      add :color, :string
+      add :rgba, :map
+      add :hsla, :map
+    end
+  end
+end