+++ test/test_helper.exs
@@ -1,1 +1,60 @@
+alias Ecto.Adapters.SQLite3
+alias KineticEcto.TestRepo
+
+# Code between the arrows is from elixir-sqlite/ecto_sqlite3: ⬇️
+# Copyright (c) 2021 Matthew A. Johnston
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to
+# deal in the Software without restriction, including without limitation the
+# rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
+# sell copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+# IN THE SOFTWARE.
+
+Logger.configure(level: :info)
+
+Application.put_env(:kinetic_ecto, TestRepo,
+  adapter: Ecto.Adapters.SQLite3,
+  database: "/tmp/kinetic_ecto_test_#{Ecto.UUID.generate()}.db",
+  pool: Ecto.Adapters.SQL.Sandbox,
+  show_sensitive_data_on_connection_error: true
+)
+
+defmodule KineticEcto.SqliteCase do
+  @moduledoc false
+
+  use ExUnit.CaseTemplate
+
+  alias Ecto.Adapters.SQL.Sandbox
+
+  setup do
+    :ok = Sandbox.checkout(TestRepo)
+    # on_exit(fn -> Ecto.Adapters.SQL.Sandbox.checkin(TestRepo) end)
+  end
+end
+
+{:ok, _} = SQLite3.ensure_all_started(TestRepo.config(), :temporary)
+
+_ = SQLite3.storage_down(TestRepo.config())
+:ok = SQLite3.storage_up(TestRepo.config())
+
+{:ok, _} = TestRepo.start_link()
+
+:ok = Ecto.Migrator.up(TestRepo, 0, KineticEcto.TestMigration, log: false)
+Ecto.Adapters.SQL.Sandbox.mode(TestRepo, :manual)
+Process.flag(:trap_exit, true)
+
+# Code between the arrows is from elixir-sqlite/ecto_sqlite3: ⬆️
+
 ExUnit.start()