+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # KineticEcto Changelog
 
+## 1.1.0 / 2025-01-29
+
+- Added `KineticEcto.Transact`, an easy way to add Saša Juriċ's
+  `Repo.transact/2` to your arsenal.
+
 ## 1.0.0 / 2025-01-19
 
 - Initial release, including changeset validation extensions